refactor: moved scalability analysis tests to profiling
